only allow one internal dolt server
